No more crash when exiting "Component size had to be atleast 8bytes in 64bit"

+++ RootSystems/Include/Shadowcaster.h
@@ -9,6 +9,6 @@ namespace RootForce
 	{
 		Shadowcaster() : m_directionalLightSlot(0) {}
 
-		int m_directionalLightSlot;
+		long long m_directionalLightSlot;
 	};
 }
